Rename pds stream/ to event-stream/
