+++ packages/pds/src/event-stream/consumers/remove-member.ts
