+++ packages/pds/src/index.ts
@@ -12,14 +12,14 @@ import { DidResolver } from '@atproto/did-resolver'
 import API, { health } from './api'
 import Database from './db'
 import ServerAuth from './auth'
-import * as streamConsumers from './stream/consumers'
+import * as streamConsumers from './event-stream/consumers'
 import * as error from './error'
 import { httpLogger, loggerMiddleware } from './logger'
 import { ServerConfig, ServerConfigValues } from './config'
 import { Locals } from './locals'
 import { ServerMailer } from './mailer'
 import { createTransport } from 'nodemailer'
-import SqlMessageQueue from './stream/message-queue'
+import SqlMessageQueue from './event-stream/message-queue'
 import { ImageUriBuilder } from './image/uri'
 import { BlobDiskCache, ImageProcessingServer } from './image/server'
 import { createServices } from './services'
