+++ packages/pds/src/services/record/plugins/assertion.ts
@@ -3,8 +3,8 @@ import { AtUri } from '@atproto/uri'
 import { CID } from 'multiformats/cid'
 import * as Assertion from '../../../lexicon/types/app/bsky/graph/assertion'
 import * as lex from '../../../lexicon/lexicons'
-import * as messages from '../../../stream/messages'
-import { Message } from '../../../stream/messages'
+import * as messages from '../../../event-stream/messages'
+import { Message } from '../../../event-stream/messages'
 import DatabaseSchema from '../../../db/database-schema'
 import RecordProcessor from '../processor'
