+++ packages/pds/src/services/record/plugins/confirmation.ts
@@ -3,8 +3,8 @@ import { AtUri } from '@atproto/uri'
 import { CID } from 'multiformats/cid'
 import * as Confirmation from '../../../lexicon/types/app/bsky/graph/confirmation'
 import * as lex from '../../../lexicon/lexicons'
-import * as messages from '../../../stream/messages'
-import { Message } from '../../../stream/messages'
+import * as messages from '../../../event-stream/messages'
+import { Message } from '../../../event-stream/messages'
 import { APP_BSKY_GRAPH } from '../../../lexicon'
 import DatabaseSchema from '../../../db/database-schema'
 import RecordProcessor from '../processor'
