+++ packages/pds/src/services/record/plugins/declaration.ts
@@ -3,7 +3,7 @@ import { AtUri } from '@atproto/uri'
 import { CID } from 'multiformats/cid'
 import * as Declaration from '../../../lexicon/types/app/bsky/system/declaration'
 import * as lex from '../../../lexicon/lexicons'
-import { Message } from '../../../stream/messages'
+import { Message } from '../../../event-stream/messages'
 import DatabaseSchema from '../../../db/database-schema'
 import RecordProcessor from '../processor'
