+++ packages/pds/src/services/record/plugins/post.ts
@@ -5,8 +5,8 @@ import { Record as PostRecord } from '../../../lexicon/types/app/bsky/feed/post'
 import { Main as ImagesEmbedFragment } from '../../../lexicon/types/app/bsky/embed/images'
 import { Main as ExternalEmbedFragment } from '../../../lexicon/types/app/bsky/embed/external'
 import * as lex from '../../../lexicon/lexicons'
-import * as messages from '../../../stream/messages'
-import { Message } from '../../../stream/messages'
+import * as messages from '../../../event-stream/messages'
+import { Message } from '../../../event-stream/messages'
 import DatabaseSchema from '../../../db/database-schema'
 import RecordProcessor from '../processor'
