+++ packages/pds/src/services/record/processor.ts
@@ -3,7 +3,7 @@ import { CID } from 'multiformats/cid'
 import { AtUri } from '@atproto/uri'
 import * as common from '@atproto/common'
 import DatabaseSchema from '../../db/database-schema'
-import { Message } from '../../stream/messages'
+import { Message } from '../../event-stream/messages'
 import { lexicons } from '../../lexicon/lexicons'
 
 type RecordProcessorParams<T, S> = {
