+++ packages/pds/src/services/repo/index.ts
@@ -4,7 +4,7 @@ import { BlobStore, Repo } from '@atproto/repo'
 import { InvalidRequestError } from '@atproto/xrpc-server'
 import Database from '../../db'
 import { dbLogger as log } from '../../logger'
-import { MessageQueue } from '../../stream/types'
+import { MessageQueue } from '../../event-stream/types'
 import SqlBlockstore from '../../sql-blockstore'
 import { PreparedCreate, PreparedWrites } from '../../repo/types'
 import { RecordService } from '../record'
