+++ packages/pds/tests/seeds/basic.ts
@@ -1,4 +1,4 @@
-import { MessageQueue } from '../../src/stream/types'
+import { MessageQueue } from '../../src/event-stream/types'
 import { SeedClient } from './client'
 import scenesSeed from './scenes'
 import usersSeed from './users'
